TimerEvent 작성

+++ EmberFallServer/GridWorld.cpp
@@ -4,8 +4,8 @@
 
 GridWorld::GridWorld(float cellWidth, float cellHeight, float mapWidth, float mapHeight, SimpleMath::Vector2 center) 
     : mGridWidth{ cellWidth }, mGridHeight{ cellHeight } {
-    mWidth = std::ceil(mapWidth / cellWidth);
-    mHeight = std::ceil(mapHeight / cellHeight);
+    mWidth = static_cast<INT32>(std::ceil(mapWidth / cellWidth));
+    mHeight = static_cast<INT32>(std::ceil(mapHeight / cellHeight));
     SimpleMath::Vector2 mapLT{ -mapWidth / 2.0f, -mapHeight / 2.0f };
 
     for (size_t y = 0; y < mWidth; ++y) {
@@ -29,8 +29,8 @@ void GridWorld::Update(const std::vector<std::shared_ptr<class GameObject>>& obj
 
     for (auto& object : objects) {
         decltype(auto) obb = std::static_pointer_cast<OrientedBoxCollider>(object->GetCollider())->GetBoundingBox();
-        auto aabbMin = obb.Center - (SimpleMath::Vector3{ obb.Extents } * 1.415); // root 2 (OBB를 감싸는 최소한의 AABB)
-        auto aabbMax = obb.Center + (SimpleMath::Vector3{ obb.Extents } * 1.415);
+        auto aabbMin = obb.Center - (SimpleMath::Vector3{ obb.Extents } * 1.415f); // root 2 (OBB를 감싸는 최소한의 AABB)
+        auto aabbMax = obb.Center + (SimpleMath::Vector3{ obb.Extents } * 1.415f);
 
         INT32 minX = std::max(static_cast<INT32>(aabbMin.x / mGridWidth), 0);
         INT32 maxX = std::min(static_cast<INT32>(aabbMax.x / mGridWidth), mWidth);
